Simplify page structure and fiddle with CSS
Attempt to fix weird Chrome behaviour when following a link to a
footnote or other anchor near the bottom of the page. Unlike Firefox,
Chrome scrolls the entire page up off the top of the screen, changing
window.pageYOffset, which it shouldn't. I've no idea why, but this
improves things a little - you can at least recover now.